Only use errtrace in shells that support it
Using the -E / errtrace option from the bash builtins breakes roundup
on shells that do not support it, eg. busyboxs' sh.
The same goes for the bash specific ERR signal which is not POSIX
compliant. Both should only be used if roundup is executed in a shell
that support it.